requirements.txt
attrs==21.4.0
Automat==20.2.0
autopep8==1.6.0
beautifulsoup4==4.12.2
cffi==1.15.0
constantly==15.1.0
cryptography==36.0.1
cycler==0.11.0
fonttools==4.28.5
google==3.0.0
googleapis-common-protos==1.62.0
grpcio==1.60.0
grpcio-tools==1.60.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
jsonrpc==3.0.1
kiwisolver==1.3.2
matplotlib==3.5.1
mypy-protobuf==3.5.0
numpy==1.22.1
packaging==21.3
Pillow==9.0.0
protobuf==4.25.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
pyOpenSSL==21.0.0
pyparsing==3.0.7
python-dateutil==2.8.2
service-identity==21.1.0
six==1.16.0
soupsieve==2.5
toml==0.10.2
Twisted==21.7.0
types-protobuf==4.24.0.4
typing_extensions==4.0.1
zope.interface==5.4.0
